Move newbranch to standalone package
